chore: Move class to service to align with infra
