+++ tests/Unit/PlaylistAuthPasswordGeneratorTest.php
@@ -1,14 +1,14 @@
 <?php
 
-use App\Support\PlaylistAuthPasswordGenerator;
+use App\Services\PasswordGeneratorService;
 
 it('generates passwords that meet the playlist auth rules', function (): void {
     foreach (range(1, 100) as $ignored) {
-        $password = PlaylistAuthPasswordGenerator::generate();
+        $password = PasswordGeneratorService::generate();
 
         expect($password)->toHaveLength(10)
             ->and($password)->toMatch('/^[a-z0-9]+$/')
-            ->and(PlaylistAuthPasswordGenerator::isValid($password))->toBeTrue();
+            ->and(PasswordGeneratorService::isValid($password))->toBeTrue();
 
         $letterCount = preg_match_all('/[a-z]/', $password);
         $numberCount = preg_match_all('/[0-9]/', $password);
@@ -20,14 +20,14 @@ it('generates passwords that meet the playlist auth rules', function (): void {
 });
 
 it('respects a custom length while keeping the same character rules', function (): void {
-    $password = PlaylistAuthPasswordGenerator::generate(16);
+    $password = PasswordGeneratorService::generate(16);
 
     expect($password)->toHaveLength(16)
-        ->and(PlaylistAuthPasswordGenerator::isValid($password))->toBeTrue();
+        ->and(PasswordGeneratorService::isValid($password))->toBeTrue();
 });
 
 it('rejects passwords that break the character mix rules', function (string $password): void {
-    expect(PlaylistAuthPasswordGenerator::isValid($password))->toBeFalse();
+    expect(PasswordGeneratorService::isValid($password))->toBeFalse();
 })->with([
     'too short' => ['abc12'],
     'uppercase letters' => ['abcdefghA1'],
